# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpsPilot API
  description: |
    OpsPilot - AI-Powered Infrastructure Operations Platform API.
    All protected endpoints require a Bearer JWT token in the Authorization header.
  version: 2.0.0
  contact:
    name: OpsPilot Team
servers:
  - url: http://localhost:3001
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      responses:
        '200':
          description: List of jobs
    post:
      tags: [Jobs]
      summary: Create job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Job created
  /api/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details
    delete:
      tags: [Jobs]
      summary: Delete job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Job deleted
  /api/jobs/{id}/execute:
    post:
      tags: [Jobs]
      summary: Execute job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Job execution started
  /api/jobs/{id}/runs:
    get:
      tags: [Jobs]
      summary: List job runs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of job runs
  /api/jobs/runs/{run_id}:
    get:
      tags: [Jobs]
      summary: Get job run detail
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run details with steps